Audio: copier: move copier related header to copier folder
These headers mostly only used by copier itself, move to
copier folder will make this folder cover all related code.
