+++ src/audio/copier/copier_dai.c
@@ -7,9 +7,9 @@
 #include <sof/trace/trace.h>
 #include <sof/audio/component_ext.h>
 #include <ipc4/copier.h>
-#include <sof/audio/dai_copier.h>
 #include <ipc/dai.h>
 #include <sof/audio/module_adapter/module/generic.h>
+#include "dai_copier.h"
 
 LOG_MODULE_DECLARE(copier, CONFIG_SOF_LOG_LEVEL);
