+++ src/audio/copier/copier_host.c
@@ -5,11 +5,11 @@
 // Author: Baofeng Tian <baofeng.tian@intel.com>
 
 #include <sof/audio/component_ext.h>
-#include <sof/audio/host_copier.h>
 #include <sof/tlv.h>
 #include <sof/trace/trace.h>
 #include <ipc4/copier.h>
 #include <sof/audio/module_adapter/module/generic.h>
+#include "host_copier.h"
 
 LOG_MODULE_DECLARE(copier, CONFIG_SOF_LOG_LEVEL);
