+++ src/audio/copier/copier_ipcgtw.c
@@ -9,7 +9,7 @@
 #include <sof/ut.h>
 #include <rtos/init.h>
 #include <ipc4/copier.h>
-#include <sof/audio/ipcgtw_copier.h>
+#include "ipcgtw_copier.h"
 
 LOG_MODULE_REGISTER(ipcgtw, CONFIG_SOF_LOG_LEVEL);
