+++ src/audio/dai-zephyr.c
@@ -7,7 +7,6 @@
 
 #include <sof/audio/buffer.h>
 #include <sof/audio/component_ext.h>
-#include <sof/audio/dai_copier.h>
 #include <sof/audio/format.h>
 #include <sof/audio/pipeline.h>
 #include <sof/common.h>
@@ -36,6 +35,7 @@
 #include <errno.h>
 #include <stddef.h>
 #include <stdint.h>
+#include "copier/dai_copier.h"
 
 #include <zephyr/device.h>
 #include <zephyr/drivers/dai.h>
